cleaned up to move generating rust code output to specific rust module
