+++ src/generator/generators/mod.rs
@@ -1,0 +1,4 @@
+mod types;
+pub mod rust;
+
+pub use types::*;
