+++ src/main.rs
@@ -4,8 +4,9 @@ mod generator;
 mod parser;
 
 use filesystem::get_latest_stable_specifications;
-use generator::generate;
+use generator::rust::generator::RustGenerator;
 use std::time::Instant;
+use crate::generator::Generator;
 
 fn main() {
     let settings = cli::get_settings();
@@ -16,7 +17,8 @@ fn main() {
             let specifications =
                 get_latest_stable_specifications(settings.output_specification_folder.as_str());
 
-            generate(settings.output_folder.as_str(), &specifications);
+            let mut rust_generator = RustGenerator::default(); 
+            rust_generator.generate(settings.output_folder.as_str(), &specifications);
 
             let elapsed = now.elapsed();
             println!("Elapsed: {:.?}", elapsed);
